refactor(results): drop the results_sidecar re-export, import markers directly
Re-exporting the marker contract from the FastAPI sidecar left two names for
one thing and kept the sidecar looking like the owner of a contract it no
longer defines. Repoint all 28 importers in src and tests at
aiperf.common.results_markers, and reduce the sidecar to importing only the
names its own handlers use.
