+++ src/aiperf/api/routers/results.py
@@ -31,13 +31,13 @@ from aiperf.common.hooks import on_message
 from aiperf.common.messages import ProcessAllResultsMessage
 from aiperf.common.mixins.message_bus_mixin import MessageBusClientMixin
 from aiperf.common.models.record_models import ProcessRecordsResult
-from aiperf.config.defaults import OutputDefaults
 from aiperf.common.results_markers import (
     CHECKPOINTS_DIR_NAME,
     READY_MARKER_NAME,
     _is_processing,
     ready_marker_path,
 )
+from aiperf.config.defaults import OutputDefaults
 
 ResultsDep = Annotated["ResultsRouter", component_dependency("results")]
