+++ tests/unit/api/test_api_service.py
@@ -897,7 +897,7 @@ class TestResultsListEndpoint:
         """Test listing results with files in directory."""
         from unittest.mock import MagicMock
 
-        from aiperf.kubernetes.results_sidecar import write_ready_marker
+        from aiperf.common.results_markers import write_ready_marker
 
         # Create test files
         (tmp_path / "metrics.json").write_text('{"test": 1}')
@@ -932,7 +932,7 @@ class TestResultsFileEndpoints:
         """Test returns 404 for nonexistent file."""
         from unittest.mock import MagicMock
 
-        from aiperf.kubernetes.results_sidecar import write_ready_marker
+        from aiperf.common.results_markers import write_ready_marker
 
         write_ready_marker(tmp_path)
         mock_output = MagicMock()
@@ -952,7 +952,7 @@ class TestResultsFileEndpoints:
         """Test file streams content with correct headers."""
         from unittest.mock import MagicMock
 
-        from aiperf.kubernetes.results_sidecar import write_ready_marker
+        from aiperf.common.results_markers import write_ready_marker
 
         test_file = tmp_path / "profile_export.json"
         test_file.write_text('{"metrics": {"latency": 100}}')
@@ -979,7 +979,7 @@ class TestResultsFileEndpoints:
         """Test path traversal attempts are rejected."""
         from unittest.mock import MagicMock
 
-        from aiperf.kubernetes.results_sidecar import write_ready_marker
+        from aiperf.common.results_markers import write_ready_marker
 
         write_ready_marker(tmp_path)
         mock_output = MagicMock()
@@ -998,7 +998,7 @@ class TestResultsFileEndpoints:
         """Test result file endpoint supports gzip compression."""
         from unittest.mock import MagicMock
 
-        from aiperf.kubernetes.results_sidecar import write_ready_marker
+        from aiperf.common.results_markers import write_ready_marker
 
         test_file = tmp_path / "metrics.json"
         test_file.write_text('{"metrics": {"latency": 100}}')
