+++ tests/unit/operator/test_completion_record_results.py
@@ -18,8 +18,8 @@ from unittest.mock import MagicMock, patch
 
 import orjson
 
+from aiperf.common.results_markers import ready_marker_path
 from aiperf.kubernetes.crd_models import ControllerFetchResult
-from aiperf.kubernetes.results_sidecar import ready_marker_path
 from aiperf.operator.handlers.completion import _record_results_on_status
 from aiperf.operator.results_layout import epoch_key_from_body, write_latest
