+++ tests/unit/operator/test_results_analytics.py
@@ -17,7 +17,7 @@ from fastapi import FastAPI
 from fastapi.testclient import TestClient
 
 from aiperf.common.redact import REDACTED_VALUE
-from aiperf.kubernetes.results_sidecar import write_ready_marker
+from aiperf.common.results_markers import write_ready_marker
 from aiperf.operator import runs_index
 from aiperf.operator.results_db import ResultsDB
 from aiperf.operator.results_layout import run_dir, write_latest
